modern header/nav combo svd decorated
