+++ src/App.css
@@ -8,6 +8,12 @@
   padding: 0;
   line-height: 1.5rem;
 }
+.grid{
+  display: grid;
+}
+.flex {
+  display: flex;
+}
 :root {
   --from : #1B1B3A;
   --to : #1e2945;
@@ -34,16 +40,14 @@ body{
   box-shadow: 0px 1px 2px var(--color);
   animation: gradient 1s ease infinite;
 }
-canvas {
-  position: fixed;
-  top: 10vh;
-  opacity: 0.6;
-}
 .name {
   font-size: large;
 }
 .flex {
   display: flex;
+}
+.mtp{
+  margin-top: 2rem;
 }
 .vc {
   align-items: center;
@@ -133,8 +137,8 @@ fieldset {
   background: var(--to);
   font-weight: 500;
   margin-top: 3%;
-  z-index: 10;
-  opacity: 0;
+  z-index: 5;
+  opacity: 100;
 }
 legend{
   font-size: large;
@@ -144,6 +148,10 @@ legend{
 }
 .iconwrap {
   display: grid;
+  height: 100%;
+  aspect-ratio: 1;
+  border-radius: 10px;
+  border: 1px solid wheat;
   align-content: center;
   justify-items: center;
   margin:0rem 0.5rem;
@@ -167,10 +175,6 @@ a {
   display: grid;
   align-content: center;
   justify-items: center;
-}
-a:hover > .roundbtn {
-  transform: scale(1.2);
-  transition: 0.1s ease-in;
 }
 .sc {
   margin-top: 2%;
