+++ src/App.jsx
@@ -1,49 +1,45 @@
 import './App.css';
 import Loader from './Loader';
 import Navbar from './components/Navbar'
-import Main from './components/Main';
 import Defenitions from './components/Defenitions';
 import Profiles from './components/Profiles';
 import Footer from './components/Footer';
 import {useState} from 'react';
-import Particles from 'react-particles-js';
-import {config} from './AppParticleConfig';
+import mypic from './components/alphabgpic.png';
+
 
 function App() {
 
-  let darkState = !(localStorage.getItem('theme') === 'false');
+  // let darkState = !(localStorage.getItem('theme') === 'false');
 
-    if(!darkState)
-    {
-        document.querySelector(":root").style.setProperty("--from","#d4f3ff");
-        document.querySelector(":root").style.setProperty("--to","#97ffc5");
-        document.querySelector(":root").style.setProperty("--color","#000000");
+  //   if(!darkState)
+  //   {
+  //       document.querySelector(":root").style.setProperty("--from","#d4f3ff");
+  //       document.querySelector(":root").style.setProperty("--to","#97ffc5");
+  //       document.querySelector(":root").style.setProperty("--color","#000000");
+  //   }
+
+    const [load,setLoad] = useState(false);
+    
+    const pic = new Image();
+    pic.src = mypic;
+    pic.onload = () => {
+        setLoad(true)
     }
-
-  const [load,setLoad] = useState(false);
-  const coverUrl = new Image();
-  coverUrl.src = "https://images.hdqwalls.com/wallpapers/linux-programmer-pixel-art-4k-ld.jpg";
-  const [start, setStart] = useState(false);
-  coverUrl.onload = () => {
-      setLoad(true)
-  }
-  console.log(darkState);
-  return (
-    <>
+    return (
+      <>
       {
-        load? <ShowContent appearState={start} appearMethod={setStart} mainCover={coverUrl.src} theme={darkState}/> : <Loader />
+        load?<ShowContent pic={setLoad}/>:<Loader />
       }
-    </>
-  );
+      </>
+    );
 }
 const ShowContent = (props) => {
   return (
     <>
-      <Navbar themeAccent={props.theme}/>
-      <Particles className="particles" params={config} height="60vh"/>
-        <Main appearState={props.appearState} appearMethod={props.appearMethod} coverPhoto={props.mainCover}/>
-        <Defenitions appearState={props.appearState} />
-        <Profiles appearState={props.appearState}/>
+        <Navbar pic={props.pic}/>
+        <Defenitions />
+        <Profiles />
         <Footer/>
     </>
   )
