+++ src/components/Defenitions.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CSSTransition } from 'react-transition-group';
+
 import Fields from './Fields';
 import { lang } from './languages';
 
@@ -12,14 +12,11 @@ const Tab = () => {
 export default function Defenitions(props) {
     return (
         <div className="fill flex vc hc">
-        <CSSTransition in={props.appearState} appear={false} timeout={900} classNames="fieldset">
-
             <fieldset className="fieldset">
                 <legend align="center"> {"<"}PERSONAL INFO{">"} </legend>
                 <Fields field="Username" value="dev-D"/> <br/><br/>
                 <Fields field="D.O.B" value="25/05/2002" /><br/><br/>
-                <b> Education :</b> <br/> <br/>
-
+                <b> Education :</b> <br/> <br/> 
                 <Tab/><Fields field="L.P School" value="St: Dominic Savio"/> <br/>
                 <Tab/><Fields field="U.P School" value="M.K.M" /> <br/>
                 <Tab/><Fields field="H.S School" value="M.K.M" /> <br/>
@@ -30,11 +27,8 @@ export default function Defenitions(props) {
                     lang.map((value,index) => {
                         return <> <Fields key={index} icon={value.icon} value={value.name} color={value.color}/></>
                     })
-                }
-
+                }   
             </fieldset>
-            
-        </CSSTransition>
         </div>
     )
 }
