+++ src/components/navbar.css
@@ -1,0 +1,92 @@
+
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300&display=swap');
+*{
+    margin: 0;
+    padding: 0;
+    color: white;
+    font-family: 'Poppins', sans-serif;
+}
+canvas {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 10vw;
+    opacity: 0.6;
+    z-index: 3;
+}
+header{
+    max-width: 100vw;
+    height: 80vh;
+    background: rgb(23,22,42);
+    background: linear-gradient(90deg, rgba(23,22,42,1) 0%, rgba(47,25,94,1) 35%, rgba(83,129,205,1) 100%);
+    display: grid;
+    justify-items: center;
+    clip-path: polygon(0% 0%, 100% 0%, 100% 88%, 50% 100%, 0 88%);
+    overflow: hidden;
+}
+nav{
+    width: 90% ;
+    height: 6vh;
+    margin-top: 1rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+ul{
+    display: flex;
+    list-style: none;
+}
+li{
+    padding: 1rem;
+}
+.segment{
+    width: 98vw;
+    height: 65vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.desc{
+    width: 40%;
+    height: 100%;
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    overflow-y: hidden;
+}
+.svg{
+    width: 80%;
+    transition: 0.5s ease-out;
+    z-index: 4;
+}
+.pic {
+    position: absolute;
+    bottom: -0.5;
+    height: 110%;
+    transition: 0.6s ease-out;
+    z-index: 4;
+}
+.desc:hover > .pic{
+    transform: scale(1.05) translateX(-10%);
+}
+.desc:hover > .svg{
+    transform: translateY(-4%) scale(1.05);
+}
+.segment2{
+    max-width: 100vw;
+    align-items: center;
+    justify-content: center;
+}
+.caption{
+    width: 60%;
+    height: 6rem;
+    color: black;
+    background: white;
+    align-items: center;
+    justify-content: center;
+    border-radius: 10px;
+    box-shadow: 1px 1px 5px 2px rgba(0, 0, 0, 0.576);
+    transform: translateY(-30%);
+    z-index: 5;
+}
